fix(shared): make :shared Android-only, fix imports and missing files
- :shared's commonMain depended on androidx.navigation3:navigation3-ui
  and androidx.lifecycle:lifecycle-viewmodel-navigation3, both
  Android-only AndroidX artifacts with no iOS klib. Once base_kmp_module
  added iosArm64/iosSimulatorArm64 targets (Phase 11), this became a
  hard "KMP Dependencies Resolution Failure" that poisoned the whole
  commonMain compile - manifesting as mass "unresolved reference" noise
  for completely unrelated symbols (Firebase, koinViewModel, Res.string.*)
  since the classpath itself never resolved. The composition root here
  is 100% Android Compose code with no iOS entry point yet, so move the
  whole source set from commonMain to androidMain - honest, matches the
  "not attempted blind" stance already used for the other Phase 11 iOS
  stubs.
- MainViewModel.kt needed libs.gitlive.firebase.auth (checks
  Firebase.auth.currentUser directly), missing from shared/build.gradle.kts.
- Same recurring fixes as prior phases: wildcard-import generated
  resources (BottomNavDestination.kt + 6 onboarding screens), wrong
  org.koin.androidx.compose.koinViewModel import (3 files, needs
  org.koin.compose.viewmodel), kotlinx-datetime Month.number ->
  monthNumber (AppSemesterHistoryScreen.kt).
- shared/consumer-rules.pro was missing entirely - base_kmp_module
  always declares consumerProguardFiles("consumer-rules.pro").
