+++ shared/consumer-rules.pro
@@ -1,0 +1,2 @@
+# Consumer ProGuard rules for shared module - no rules needed; it's only consumed by :app,
+# which already keeps what it needs via its own R8 configuration.
