+++ shared/src/androidMain/kotlin/com/hussienfahmy/myGpaManager/navigation/screens/AppSemesterDetailScreen.kt
