+++ shared/src/androidMain/kotlin/com/hussienfahmy/myGpaManager/navigation/screens/AppSubjectSettingsScreen.kt
